@@ -1,11 +1,12 @@
+using System.Collections.Generic;
+using Microsoft.AspNetCore.Identity;
+
 namespace AdmonSoptec.API.Models
 {
-    public class User
-    {
-      public int Id { get; set; }  
-      public string  Email { get; set; }
-      public string  Username { get; set; }
-      public byte[] PaswswordHash { get; set; }
-     public byte[] PaswswordSalt { get; set; }
+    public class User: IdentityUser<int>
+    {      public string  Telefono { get; set; }
+      public string  Puesto { get; set; }
+      public virtual ICollection<UserRole> UserRoles { get; set; }
     }
+
 }
